wait for file to be written in JsonStorageTest
fixes #3854
