+++ bundles/test/org.eclipse.smarthome.test/src/main/java/org/eclipse/smarthome/test/java/JavaOSGiTest.java
@@ -16,10 +16,7 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
-import java.util.function.BooleanSupplier;
 import java.util.function.Predicate;
-import java.util.function.Supplier;
 
 import org.eclipse.smarthome.core.autoupdate.AutoUpdateBindingConfigProvider;
 import org.eclipse.smarthome.test.OSGiTest;
@@ -41,10 +38,7 @@ import org.osgi.framework.ServiceRegistration;
  *
  * @author Markus Rathgeb - Create a pure Java implementation based on the Groovy {@link OSGiTest} class
  */
-public class JavaOSGiTest {
-
-    protected final int DFL_TIMEOUT = 10000;
-    protected final int DFL_SLEEP_TIME = 50;
+public class JavaOSGiTest extends JavaTest {
 
     private final Map<String, ServiceRegistration<?>> registeredServices = new HashMap<>();
     protected BundleContext bundleContext;
@@ -265,133 +259,6 @@ public class JavaOSGiTest {
     }
 
     /**
-     * Wait until the condition is fulfilled or the timeout is reached.
-     *
-     * <p>
-     * This method uses the default timing parameters.
-     *
-     * @param condition the condition to check
-     * @return true on success, false on timeout
-     */
-    protected boolean waitFor(BooleanSupplier condition) {
-        return waitFor(condition, DFL_TIMEOUT, DFL_SLEEP_TIME);
-    }
-
-    /**
-     * Wait until the condition is fulfilled or the timeout is reached.
-     *
-     * @param condition the condition to check
-     * @param timeout timeout
-     * @param sleepTime interval for checking the condition
-     * @return true on success, false on timeout
-     */
-    protected boolean waitFor(BooleanSupplier condition, int timeout, int sleepTime) {
-        int waitingTime = 0;
-        boolean rv;
-        while (!(rv = condition.getAsBoolean()) && waitingTime < timeout) {
-            waitingTime += sleepTime;
-            internalSleep(sleepTime);
-        }
-        return rv;
-    }
-
-    /**
-     * Wait until the assertion is fulfilled or the timeout is reached.
-     *
-     * <p>
-     * This method uses the default timing parameters.
-     *
-     * @param assertion closure that must not have an argument
-     */
-    protected void waitForAssert(Runnable assertion) {
-        waitForAssert(assertion, null, DFL_TIMEOUT, DFL_SLEEP_TIME);
-    }
-
-    /**
-     * Wait until the assertion is fulfilled or the timeout is reached.
-     *
-     * @param assertion the logic to execute
-     * @param timeout timeout
-     * @param sleepTime interval for checking the condition
-     */
-    protected void waitForAssert(Runnable assertion, int timeout, int sleepTime) {
-        waitForAssert(assertion, null, timeout, sleepTime);
-    }
-
-    /**
-     * Wait until the assertion is fulfilled or the timeout is reached.
-     *
-     * <p>
-     * This method uses the default timing parameters.
-     *
-     * @param assertion the logic to execute
-     * @return the return value of the supplied assertion object's function on success
-     */
-    protected <T> T waitForAssert(Supplier<T> assertion) {
-        return waitForAssert(assertion, null, DFL_TIMEOUT, DFL_SLEEP_TIME);
-    }
-
-    /**
-     * Wait until the assertion is fulfilled or the timeout is reached.
-     *
-     * @param assertion the logic to execute
-     * @param timeout timeout
-     * @param sleepTime interval for checking the condition
-     * @return the return value of the supplied assertion object's function on success
-     */
-    protected <T> T waitForAssert(Supplier<T> assertion, int timeout, int sleepTime) {
-        return waitForAssert(assertion, null, timeout, sleepTime);
-    }
-
-    /**
-     * Wait until the assertion is fulfilled or the timeout is reached.
-     *
-     * @param assertion the logic to execute
-     * @param beforeLastCall logic to execute in front of the last call to ${code assertion}
-     * @param sleepTime interval for checking the condition
-     */
-    protected void waitForAssert(Runnable assertion, Runnable beforeLastCall, int timeout, int sleepTime) {
-        int waitingTime = 0;
-        while (waitingTime < timeout) {
-            try {
-                assertion.run();
-                return;
-            } catch (final Error | NullPointerException error) {
-                waitingTime += sleepTime;
-                internalSleep(sleepTime);
-            }
-        }
-        if (beforeLastCall != null) {
-            beforeLastCall.run();
-        }
-        assertion.run();
-    }
-
-    /**
-     * Wait until the assertion is fulfilled or the timeout is reached.
-     *
-     * @param assertion the logic to execute
-     * @param beforeLastCall logic to execute in front of the last call to ${code assertion}
-     * @param sleepTime interval for checking the condition
-     * @return the return value of the supplied assertion object's function on success
-     */
-    protected <T> T waitForAssert(Supplier<T> assertion, Runnable beforeLastCall, long timeout, int sleepTime) {
-        final long timeoutNs = TimeUnit.MILLISECONDS.toNanos(timeout);
-        final long startingTime = System.nanoTime();
-        while (System.nanoTime() - startingTime < timeoutNs) {
-            try {
-                return assertion.get();
-            } catch (final Error | NullPointerException error) {
-                internalSleep(sleepTime);
-            }
-        }
-        if (beforeLastCall != null) {
-            beforeLastCall.run();
-        }
-        return assertion.get();
-    }
-
-    /**
      * Returns the interface name for a given service object by choosing the first interface.
      *
      * @param service service object
@@ -443,14 +310,6 @@ public class JavaOSGiTest {
                 return true;
             }
         });
-    }
-
-    private void internalSleep(long millis) {
-        try {
-            Thread.sleep(millis);
-        } catch (InterruptedException e) {
-            throw new Error("We shouldn't be interrupted while testing");
-        }
     }
 
 }
